+++ apps/hubble/src/network/sync/syncEngine.ts
@@ -53,10 +53,13 @@ const SYNC_THRESHOLD_IN_SECONDS = 10;
 // The maximum number of nodes to enque in the work queue
 const MAX_WORK_QUEUE_SIZE = 100_000;
 
+export const FIRST_SYNC_DELAY = 10 * 1000; // How long to wait after startup to start syncing
 const SYNC_MAX_DURATION = 110 * 60 * 1000; // 110 minutes, just slightly less than the periodic sync job frequency
 
-// number of CPUs, clamped between 2 and 8
-const SYNC_PARALLELISM = Math.max(Math.min(os.cpus().length, 8), 2);
+// number of CPUs, clamped between 2 and 4
+// This should be clamped between 2 and 16, but because of a perf bug serving the sync trie
+// we are limiting it to 4 for now. Once this release is rolled out, we can increase it to 16.
+const SYNC_PARALLELISM = Math.max(Math.min(os.cpus().length, 4), 2);
 const SYNC_INTERRUPT_TIMEOUT = 30 * 1000; // 30 seconds
 
 // If a peer returns over 20 timeouts during sync, we abandon syncing with them
@@ -621,7 +624,7 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
 
       // Pick a more random peers, one per sync thread. It's OK if we pick the same twice, it might
       // happen if we don't have that many peers.
-      for (let i = 0; i < SYNC_PARALLELISM; i++) {
+      for (let i = 0; i < Math.min(SYNC_PARALLELISM, peers.length); i++) {
         const randomPeer = peers[Math.floor(Math.random() * peers.length)] as PeerContact;
         secondaryContacts.push(randomPeer);
       }
@@ -1198,6 +1201,9 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
   }
 
   async processSyncWorkQueue(): Promise<void> {
+    // Cap the sync parallelism to the number of secondary rpc clients we could get
+    const syncParallelism = Math.min(SYNC_PARALLELISM, this.curSync.secondaryRpcClients.length);
+
     while (this.curSync.workQueue.length > 0) {
       // Go over the work queue, and count how many items are in progress, and how many are completed
       const newWorkQueue: SyncEngineWorkItem[] = [];
@@ -1231,8 +1237,8 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
       this.curSync.workQueue = newWorkQueue;
 
       // If there is an opputunity to do more work, we will do it
-      if (this.curSync.executingWorkCount < SYNC_PARALLELISM) {
-        const n = Math.max(0, SYNC_PARALLELISM - this.curSync.executingWorkCount);
+      if (this.curSync.executingWorkCount < syncParallelism) {
+        const n = Math.max(0, syncParallelism - this.curSync.executingWorkCount);
         // Find the "n" best work items, and start them
         const bestWorkItems = scheduledItems.sort((a, b) => b.score - a.score).slice(0, n);
 
@@ -1518,6 +1524,10 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
 
     const theirSnapshot = peerStateResult.value;
     return this.syncStatus(peerId, theirSnapshot);
+  }
+
+  public get shouldCompactDb(): boolean {
+    return this._messagesSinceLastCompaction > COMPACTION_THRESHOLD;
   }
 
   public async getSnapshot(prefix?: Uint8Array): HubAsyncResult<TrieSnapshot> {