+++ packages/shuttle/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @farcaster/hub-shuttle
 
+## 0.2.3
+
+### Patch Changes
+
+- feat(shuttle): add message state to handler
+
+## 0.2.2
+
+### Patch Changes
+
+- fix: ack the correct group name to fix stale bug
+
+## 0.2.1
+
+### Patch Changes
+
+- feat(shuttle): Make event stream more customizable
+
+## 0.2.0
+
+### Minor Changes
+
+- chore: Rename to shuttle and release beta
+
+## 0.1.6
+
+### Patch Changes
+
+- feat(shuttle): support prune, revokes and deletes
+- 1051b3dd: feat: Support redis streams for scalability and add reconciliation workers
+
 ## 0.1.5
 
 ### Patch Changes